Refactor EvalMLAnalysis modules and organize imports
Moved EvalMLAnalysis and EvalMLAnalysisV2 classes into separate dedicated modules (analysis_comp and analysis_comp_v2). Renamed analysis.py to utils.py for better clarity. Updated imports across the codebase to reflect these structural changes.
