+++ tabular_orchestrated/evalml/__init__.py
@@ -1,4 +1,5 @@
-from tabular_orchestrated.evalml.analysis import EvalMLAnalysis, EvalMLAnalysisV2
+from tabular_orchestrated.evalml.analysis.analysis_comp import EvalMLAnalysis
+from tabular_orchestrated.evalml.analysis.analysis_comp_v2 import EvalMLAnalysisV2
 from tabular_orchestrated.evalml.pipeline import EvalMLPredict, EvalMLSelectPipeline
 from tabular_orchestrated.evalml.search import EvalMLSearch
