+++ tabular_orchestrated/evalml/analysis/utils.py
@@ -52,16 +52,12 @@ from evalml.pipelines import (
     RegressionPipeline,
 )
 from evalml.problem_types import detect_problem_type
-from ml_orchestrator import artifacts
-from ml_orchestrator.artifacts import Input, Output
-
-from tabular_orchestrated.evalml.evalml import EvalMLComp
 
 EvalMLModel = Union[PipelineBase, BinaryClassificationPipeline, RegressionPipeline, MulticlassClassificationPipeline]
 
 
 @dataclasses.dataclass
-class EvalMLAnalysisUtils(EvalMLComp):
+class EvalMLAnalysisUtils:
     @staticmethod
     def detect_problem_type(y: pd.Series) -> str:
         pt = detect_problem_type(y)
@@ -155,79 +151,3 @@ class EvalMLAnalysisUtils(EvalMLComp):
             confusion = confusion_matrix(labels, y_pred)
             charts.append(confusion)
         return charts
-
-
-@dataclasses.dataclass
-class EvalMLAnalysis(EvalMLComp):
-    model: Input[artifacts.Model]
-    test_dataset: Input[artifacts.Dataset]
-    analysis: Output[artifacts.HTML]
-    metrics: Output[artifacts.Metrics]
-
-    @property
-    def problem_type(self) -> str:
-        return self.model.metadata["problem_type"]
-
-    def execute(self) -> None:
-        model = self.load_model(self.model)
-        test_df = self.load_df(self.test_dataset)
-        self.analyze(model, test_df)
-
-    def analyze(self, model, test_df):
-        y_pred = model.predict(test_df[self.model_columns(test_df)])
-        y_pred_proba = (
-            model.predict_proba(test_df[self.model_columns(test_df)]) if self.problem_type != "regression" else None
-        )
-        labels = test_df[self.target_column]
-        str_charts = self.create_charts(model, labels, y_pred, y_pred_proba)
-        metrics = EvalMLAnalysisUtils.create_metrics(labels, y_pred, y_pred_proba)
-        for metric_name, metric_value in metrics.items():
-            self.metrics.log_metric(metric_name, metric_value)
-        html_str = "<br>".join(str_charts)
-        self.save_html(self.analysis, html_str)
-        self.analysis.metadata["number of charts"] = len(str_charts)
-
-    def create_charts(self, model: EvalMLModel, labels: pd.Series, y_pred, y_pred_proba) -> list[str]:
-        charts = EvalMLAnalysisUtils.create_metric_charts(labels, y_pred, y_pred_proba)
-
-        charts.append(model.graph_feature_importance())
-        str_charts = [chart.to_html() for chart in charts]
-        return str_charts
-
-
-@dataclasses.dataclass
-class EvalMLAnalysisV2(EvalMLComp):
-    predictions: Input[artifacts.Dataset]
-    analysis: Output[artifacts.HTML]
-    metrics: Output[artifacts.Metrics]
-    pred_column: str = "pred_column"
-    proba_column_prefix: str = "proba_column"
-
-    @property
-    def problem_type(self) -> str:
-        return self.predictions.metadata["problem_type"]
-
-    def execute(self) -> None:
-        predictions_df = self.load_df(self.predictions)
-        self.analyze(predictions_df)
-
-    def analyze(self, predictions_df):
-        y_pred = predictions_df[self.pred_column]
-        y_pred_proba = None
-        if self.detect_problem_type(predictions_df[self.target_column]) != "regression":
-            y_pred_proba = EvalMLAnalysisUtils.get_proba(predictions_df, self.proba_column_prefix)
-
-        labels = predictions_df[self.target_column]
-        str_charts = self.create_charts(labels, y_pred, y_pred_proba)
-        metrics = EvalMLAnalysisUtils.create_metrics(labels, y_pred, y_pred_proba)
-        for metric_name, metric_value in metrics.items():
-            self.metrics.log_metric(metric_name, metric_value)
-        html_str = "<br>".join(str_charts)
-        self.save_html(self.analysis, html_str)
-        self.analysis.metadata["number of charts"] = len(str_charts)
-
-    def create_charts(self, labels: pd.Series, y_pred, y_pred_proba) -> list[str]:
-        charts = EvalMLAnalysisUtils.create_metric_charts(labels, y_pred, y_pred_proba)
-
-        str_charts = [chart.to_html() for chart in charts]
-        return str_charts
